third_party/blink/web_tests/animations/web-animations/element-animate-keyframe-value-warning-expected.txt

CONSOLE WARNING: Invalid keyframe value for property offsetPath: path("invalid")
CONSOLE WARNING: Invalid keyframe value for property --x: invalid )
CONSOLE WARNING: Invalid keyframe value for property --x: invalid !important
This test passes if there are three warnings about invalid keyframe values.
